+++ kmserver-core/src/types/config.d.ts
@@ -16,7 +16,6 @@ export interface Config {
 	},
 	Remote: {
 		Enabled: boolean
-		BaseHost: string
 		FrontendRoot: string
 		FrontendBundlesURL: string
 	},